ebook-agent/src/requirements.txt
llama-index>=0.10.0
llama-index-vector-stores-qdrant>=0.2.0
qdrant-client>=1.7.0
openai>=1.0.0
rich>=13.0.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
PyYAML>=6.0.0
